backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
apscheduler==3.10.4
feedparser==6.0.11
httpx==0.27.2
openai>=1.0.0
pydantic==2.10.3
sqlalchemy==2.0.36
python-dotenv==1.0.1
